apps/agentic_rag/requirements.txt
docling
langchain-openai
langchain-oracledb
chromadb
python-dotenv
openai
pydantic
fastapi
uvicorn
python-multipart
torch
pyyaml
trafilatura
gradio>=6.7.0,<7.0
open-webui
lxml_html_clean
langchain
gitingest
ollama
oracledb
sentence-transformers
pytest
pytest-asyncio
agent-reasoning
